+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Add
+- Add the shortcut '[author]' to create a link to an author
+### Change
+- Move 'author' link to '/autore/'
+### Polish
+- Add Logout icon
+
 ## [1.10.1] 2026-01-30
 ### Fix
 - Move the 'cr=disabled' query param at the end of the URL in the LS shortcut.